🏗️ Architect: Extract caching logic to CacheMixin
🏗️ Design Change: Extracted caching state and lookup logic from `GenericListGetEndpoint` into a dedicated `CacheMixin`.
♻️ DRY Gains: Removes duplication of cache methods if other base classes need caching. Decouples endpoint structure from caching state.
🛡️ Solidity: SRP improvement. Base generic endpoint no longer manages cache initialization.
⚠️ Breaking Changes: None. Caching behavior remains internal to endpoints.
